Moved entity views to their own folder, then added "Delete" buttons.
